% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs.r
\name{fluorescence.selection}
\alias{fluorescence.selection}
\title{Ranking genes from fluorescence selection data}
\usage{
fluorescence.selection(
  gene,
  sgrna,
  barcode,
  reference,
  fluorescence,
  selection.ratio,
  n.to.consider = NA
)
}
\arguments{
\item{gene}{vector of character, gene names.}

\item{sgrna}{vector of character, sgrna sequences or names.}

\item{barcode}{vector of character, barcode sequences or names.}

\item{reference}{vector of integer, raw counts of barcodes in
reference group.}

\item{fluorescence}{vector of integer, raw counts of barcodes in
fluorescence group.}

\item{selection.ratio}{numeric, value should be in (0, 1), the
ratio of cells in fluorescence group which were selected from
reference group.}

\item{n.to.consider}{integer, the number of barcodes to be used
for ranking of genes. Genes with more barcodes than
n.to.consider will be ranked by randomly sampled barcodes
(with bootstrapping to make stable results).}
}
\value{
returns the fs data type, which is a list containing gene
    ranks, barcode data, and analysis parameters.
}
\description{
Function for calculation of gene ranks from
    fluorescence enriched screening data.
}
\examples{
result <- fluorescence.selection(
    gene, sgrna, barcode, reference, fluorescence, selection.ratio=0.01
)

}
